Import game state into frontend and load it into the UI
A lot going on, but I think it's almost there. I think the font for the
stacksize of the players is a bit off. It's plain white which I don't
like, but that's about it.

Players are arranged around the table, conditionally rendered. It starts
with the player in the "hero" seat (bottom middle), and then begins
filling in counter clockwise. To make the hero dynamic, it might just
need a rotation of the array. That's for a later date though.